Detects existing symlinks and handles them properly.

+++ update.sh
@@ -30,9 +30,26 @@ cd $dir
 echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
+echo "Creating symlinks"
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+	# If it is a symbolic link, check it for validity
+    if [[ -L ~/.$file ]]; then
+		if ! [[ $(readlink ~/.$file) = $dir/$file  ]]; then
+			echo "Symlink is incorrect: ~/.$file, re-creating properly"
+			rm ~/.$file
+			ln -s $dir/$file ~/.$file
+		else
+			echo "Symlink already exists: ~/.$file, skipping."
+		fi
+	# It's not a symlink, so move it, and create the proper symlink
+	else
+		if [[ -f ~/.$file ]]; then
+			echo "Moving existing dotfile .$file from ~ to $olddir"
+			mv ~/.$file ~/dotfiles_old/
+	    fi
+		
+		echo "Creating symlink to $file in home directory."
+		ln -s $dir/$file ~/.$file
+	fi
 done
+echo "...done"
